Protean Typing improvements - Phase 1 (#658)
* Improve static type checking support for Pylance and mypy

- Add @dataclass_transform() to all domain element decorators so Pylance
  generates correct __init__ parameters for decorated classes
- Add TYPE_CHECKING stubs for field factories (String, Integer, Float,
  etc.) that return resolved Python types instead of FieldSpec, enabling
  Pylance to infer constructor parameter types
- Thread TypeVar through decorator chain (decorators, factory functions,
  derive_element_class) to preserve class identity through registration
- Add @overload signatures to all 16 decorator methods on Domain for
  both direct (@domain.aggregate) and parameterized (@domain.aggregate())
  calling conventions
- Add meta_ ClassVar annotation and _default_options default to
  OptionsMixin for attribute visibility in type checkers
- Fix Auto identity_function parameter type to accept Callable
- Fix ValidationError messages type from dict[str, str] to
  dict[str, list[str]] to match actual usage
- Add Config2.from_object() method referenced in documentation

* Fix 189 pyright type errors with null guards and parameter fixes

Add assert-based null guards for _get_session(), id_field(), and
Message.metadata access patterns. Rename __init_subclass__ 'subclass'
params to 'cls'. Fix return type annotations in eventing and event
store ports. All 3381 tests passing.

* Fix 37 pyright type errors: return types, assignments, and operator issues

- Make DomainObjects a StrEnum to match ClassVar[str] declarations
- Fix return type annotations (set vs List, Optional wrappers, Any for
  Annotated/Literal returns)
- Cast broker config values to int/float at extraction time
- Use Optional[vo_cls] instead of runtime vo_cls | None union syntax
- Normalize email recipient fields via convert_str_values_to_list

* Fix 40 more pyright errors: call issues, index issues, and small categories

- Eliminate reportCallIssue: add None guards for dict.get()
  with type|None keys, fix abstract method signatures for _delete()
  and as_dict(), type-ignore for SQLAlchemy/Typer typing limitations
- Eliminate reportIndexIssue: fix MemoryModel.to_entity param
  type to dict[str, Any], type-ignore __dict__ writes in entity.py
- Eliminate reportOptionalOperand: assert before ~ operator
  in elasticsearch adapter
- Eliminate reportPossiblyUnboundVariable: initialize
  position_in_snapshot before conditional block
- Add return type annotations to BaseDatabaseModel.from_entity/to_entity
  abstract methods, reducing reportIncompatibleMethodOverride (77→69)
- Fix _repositories typing from defaultdict(str) to defaultdict(dict)

* Widen public API type hints for decorator-compatible type checking

Protean's @domain.command, @domain.event, @domain.aggregate, etc.
decorators dynamically add base classes at runtime, but preserve the
original type in their overload signatures. This caused pyright to
reject decorated classes when passed to framework methods.

Widen parameter types to Any for user-facing APIs:
- Domain.process(), command_handler_for(), handlers_for()
- BaseRepository.add() and get()
- EventSourcedRepository.get()
- @handle decorator target_cls (Type[BaseEvent|BaseCommand] -> type)
- EventStore.handlers_for() and command_handler_for()

* Add "Organize by Domain Concept" pattern to documentation

Captures the screaming architecture thought process from todo/screaming-arch.md
into the project documentation: folder tree owns domain concepts, framework
carries layer metadata via decorators.

- New pattern page in docs/patterns/ with full guidance (problem, pattern,
  anti-patterns, litmus tests)
- Expanded philosophy.md "Mirror the Domain in Code" with folder structure
  principles
- Added design rationale to protean new scaffold docs in cli/new.md
- Updated nav, contents, how-do-i, and patterns index

* Extend mypy plugin with decorator base class injection and fix retry_delay_seconds type

Enhance the Protean mypy plugin to understand decorator-based class
registration (@domain.aggregate, @domain.entity, etc.) by injecting
base class symbols during type analysis. This eliminates false mypy
errors for methods like raise_(), to_dict(), and auto-injected
attributes like `id` on decorated classes.

Key technical decisions:
- Use get_customize_class_mro_hook instead of get_class_decorator_hook
  because @dataclass_transform() on Domain methods causes mypy to
  bypass the decorator hook entirely
- Copy base class symbols into the class namespace rather than modifying
  info.bases to avoid metaclass conflicts with pydantic's ModelMetaclass
- Match decorators by AST inspection (MemberExpr/CallExpr) since name
  resolution may be incomplete during MRO calculation

Plugin additions:
- DECORATOR_BASE_CLASS_MAP covering all 15 domain element decorators
- _extract_decorator_name_from_expr for AST-based decorator matching
- _inject_base_class with symbol copying from base class hierarchy
- _maybe_inject_auto_id for aggregate/entity id field injection
- PROTEAN_MYPY_DEBUG=1 environment variable for diagnostic output

Testing (53 tests):
- 6 new mypy fixture files for decorator type verification
- 12 new integration tests running mypy on decorator fixtures
- 20 new unit tests for decorator mapping, extraction, and hooks

Infrastructure:
- VS Code settings: disable Pylance type checking, configure mypy extension
- CI: add dedicated mypy plugin test step before full test suite
- Makefile: add `typecheck` target
- Documentation: type-checking guide at docs/guides/type-checking.md

* Increase coverage

* More coverage tests

+++ src/protean/server/subscription/profiles.py
@@ -225,7 +225,7 @@ class SubscriptionConfig:
         default_factory=lambda: DEFAULT_CONFIG["blocking_timeout_ms"]
     )
     max_retries: int = field(default_factory=lambda: DEFAULT_CONFIG["max_retries"])
-    retry_delay_seconds: int = field(
+    retry_delay_seconds: float = field(
         default_factory=lambda: DEFAULT_CONFIG["retry_delay_seconds"]
     )
     enable_dlq: bool = field(default_factory=lambda: DEFAULT_CONFIG["enable_dlq"])
@@ -328,7 +328,7 @@ class SubscriptionConfig:
         tick_interval: Optional[int] = None,
         blocking_timeout_ms: Optional[int] = None,
         max_retries: Optional[int] = None,
-        retry_delay_seconds: Optional[int] = None,
+        retry_delay_seconds: Optional[float] = None,
         enable_dlq: Optional[bool] = None,
         position_update_interval: Optional[int] = None,
         origin_stream: Optional[str] = None,
@@ -477,7 +477,7 @@ class SubscriptionConfig:
             ("tick_interval", int),
             ("blocking_timeout_ms", int),
             ("max_retries", int),
-            ("retry_delay_seconds", int),
+            ("retry_delay_seconds", float),
             ("enable_dlq", bool),
             ("position_update_interval", int),
             ("origin_stream", str),
